Renama module assessment to DiabetesAssessment + add view assessment in module app
