+++ app/src/main/java/com/mediscreen/app/bean/AssessmentBean.java
@@ -1,0 +1,19 @@
+package com.mediscreen.app.bean;
+
+import com.mediscreen.app.constant.RiskLevel;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class AssessmentBean {
+
+    private LocalDateTime date;
+    private int scoreTrigger;
+    private RiskLevel riskLevel;
+
+}
