+++ app/src/main/java/com/mediscreen/app/constant/Gender.java
@@ -2,5 +2,5 @@ package com.mediscreen.app.constant;
 
 public enum Gender {
     M,
-    WOMAN
+    F
 }
