+++ app/src/main/java/com/mediscreen/app/constant/RiskLevel.java
@@ -1,0 +1,9 @@
+package com.mediscreen.app.constant;
+
+public enum RiskLevel {
+    NONE,
+    BORDERLINE,
+    IN_DANGER,
+    EARLY_ONSET,
+    UNKNOWN
+}
