+++ app/src/main/java/com/mediscreen/app/controller/AssessmentController.java
@@ -1,0 +1,13 @@
+package com.mediscreen.app.controller;
+
+import lombok.AllArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/assessment")
+@AllArgsConstructor
+public class AssessmentController {
+
+
+}
