+++ app/src/main/java/com/mediscreen/app/service/IAssessmentService.java
@@ -1,0 +1,9 @@
+package com.mediscreen.app.service;
+
+import com.mediscreen.app.bean.AssessmentBean;
+
+public interface IAssessmentService {
+
+    AssessmentBean getByPatientId(Long patientId);
+
+}
