--- assessment/src/main/java/com/mediscreen/assessment/AssessmentApplication.java
@@ -1,17 +1,0 @@
-package com.mediscreen.assessment;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@SpringBootApplication
-@EnableFeignClients("com.mediscreen.assessment")
-@EnableEurekaClient
-public class AssessmentApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AssessmentApplication.class, args);
-    }
-
-}
