--- assessment/src/main/java/com/mediscreen/assessment/constant/Gender.java
@@ -1,6 +1,0 @@
-package com.mediscreen.assessment.constant;
-
-public enum Gender {
-    M,
-    F
-}
