--- assessment/src/main/java/com/mediscreen/assessment/constant/RiskLevel.java
@@ -1,9 +1,0 @@
-package com.mediscreen.assessment.constant;
-
-public enum RiskLevel {
-    NONE,
-    BORDERLINE,
-    IN_DANGER,
-    EARLY_ONSET,
-    UNKNOWN
-}
