--- assessment/src/main/java/com/mediscreen/assessment/service/IAssessmentService.java
@@ -1,9 +1,0 @@
-package com.mediscreen.assessment.service;
-
-import com.mediscreen.assessment.dto.AssessmentDTO;
-
-public interface IAssessmentService {
-
-    AssessmentDTO getAssessmentByPatient(Long patientId);
-
-}
