--- assessment/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-spring.application.name=assessment-api
-server.port=9002
-################### Eureka ##########################
-eureka.client.service-url.defaultZone=http://localhost:8761/eureka/
-################### Actuator ##########################
-management.endpoints.web.exposure.include=health,info,prometheus
-################### Actuator ##########################
-logging.level.com.mediscreen.assessment=DEBUG
-################### Whitelabel error ##########################
-server.error.whitelabel.enabled=false
