--- assessment/src/test/java/com/mediscreen/assessment/AssessmentApplicationTests.java
@@ -1,13 +1,0 @@
-package com.mediscreen.assessment;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AssessmentApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
