+++ diabetesAssessment/src/main/java/com/mediscreen/diabetesAssessment/DiabetesAssessmentApplication.java
@@ -1,0 +1,17 @@
+package com.mediscreen.diabetesAssessment;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableFeignClients("com.mediscreen.diabetesAssessment")
+@EnableEurekaClient
+public class DiabetesAssessmentApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DiabetesAssessmentApplication.class, args);
+    }
+
+}
