+++ diabetesAssessment/src/main/java/com/mediscreen/diabetesAssessment/bean/NoteBean.java
@@ -1,0 +1,22 @@
+package com.mediscreen.diabetesAssessment.bean;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class NoteBean {
+
+    private String id;
+
+    private Long patientId;
+
+    private LocalDateTime date;
+
+    private String note;
+
+}
