+++ diabetesAssessment/src/main/java/com/mediscreen/diabetesAssessment/constant/Gender.java
@@ -1,0 +1,6 @@
+package com.mediscreen.diabetesAssessment.constant;
+
+public enum Gender {
+    M,
+    F
+}
