+++ diabetesAssessment/src/main/java/com/mediscreen/diabetesAssessment/constant/RiskLevel.java
@@ -1,0 +1,9 @@
+package com.mediscreen.diabetesAssessment.constant;
+
+public enum RiskLevel {
+    NONE,
+    BORDERLINE,
+    IN_DANGER,
+    EARLY_ONSET,
+    UNKNOWN
+}
