+++ diabetesAssessment/src/main/java/com/mediscreen/diabetesAssessment/service/IDiabetesAssessmentService.java
@@ -1,0 +1,9 @@
+package com.mediscreen.diabetesAssessment.service;
+
+import com.mediscreen.diabetesAssessment.dto.DiabetesAssessmentDTO;
+
+public interface IDiabetesAssessmentService {
+
+    DiabetesAssessmentDTO getAssessmentByPatient(Long patientId);
+
+}
