+++ diabetesAssessment/src/test/java/com/mediscreen/diabetesAssessment/DiabetesAssessmentApplicationTests.java
@@ -1,0 +1,13 @@
+package com.mediscreen.diabetesAssessment;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DiabetesAssessmentApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
